+++ pkg/provider/docker/events_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/moby/moby/api/types/container"
+	"github.com/moby/moby/client"
 	"github.com/neilotoole/slogt"
 	"github.com/sablierapp/sablier/pkg/provider/docker"
 	"gotest.tools/v3/assert"
@@ -19,16 +19,16 @@ func TestDockerClassicProvider_NotifyInstanceStopped(t *testing.T) {
 	ctx, cancel := context.WithTimeout(t.Context(), 30*time.Second)
 	defer cancel()
 	dind := setupDinD(t)
-	p, err := docker.New(ctx, dind.client, slogt.New(t))
+	p, err := docker.New(ctx, dind.client, slogt.New(t), "stop")
 	assert.NilError(t, err)
 
 	c, err := dind.CreateMimic(ctx, MimicOptions{})
 	assert.NilError(t, err)
 
-	inspected, err := dind.client.ContainerInspect(ctx, c.ID)
+	inspected, err := dind.client.ContainerInspect(ctx, c.ID, client.ContainerInspectOptions{})
 	assert.NilError(t, err)
 
-	err = dind.client.ContainerStart(ctx, c.ID, container.StartOptions{})
+	_, err = dind.client.ContainerStart(ctx, c.ID, client.ContainerStartOptions{})
 	assert.NilError(t, err)
 
 	<-time.After(1 * time.Second)
@@ -36,11 +36,11 @@ func TestDockerClassicProvider_NotifyInstanceStopped(t *testing.T) {
 	waitC := make(chan string)
 	go p.NotifyInstanceStopped(ctx, waitC)
 
-	err = dind.client.ContainerStop(ctx, c.ID, container.StopOptions{})
+	_, err = dind.client.ContainerStop(ctx, c.ID, client.ContainerStopOptions{})
 	assert.NilError(t, err)
 
 	name := <-waitC
 
 	// Docker container name is prefixed with a slash, but we don't use it
-	assert.Equal(t, "/"+name, inspected.Name)
+	assert.Equal(t, "/"+name, inspected.Container.Name)
 }