+++ pkg/provider/docker/testcontainers_test.go
@@ -24,18 +24,21 @@ type MimicOptions struct {
 	Labels        map[string]string
 }
 
-func (d *dindContainer) CreateMimic(ctx context.Context, opts MimicOptions) (container.CreateResponse, error) {
+func (d *dindContainer) CreateMimic(ctx context.Context, opts MimicOptions) (client.ContainerCreateResult, error) {
 	if len(opts.Cmd) == 0 {
 		opts.Cmd = []string{"/mimic", "-running", "-running-after=1s", "-healthy=false"}
 	}
 
 	d.t.Log("Creating mimic container with options", opts)
-	return d.client.ContainerCreate(ctx, &container.Config{
-		Entrypoint:  opts.Cmd,
-		Image:       "sablierapp/mimic:v0.3.1",
-		Labels:      opts.Labels,
-		Healthcheck: opts.Healthcheck,
-	}, &container.HostConfig{RestartPolicy: opts.RestartPolicy}, nil, nil, "")
+	return d.client.ContainerCreate(ctx, client.ContainerCreateOptions{
+		Config: &container.Config{
+			Entrypoint:  opts.Cmd,
+			Image:       "sablierapp/mimic:v0.3.1",
+			Labels:      opts.Labels,
+			Healthcheck: opts.Healthcheck,
+		},
+		HostConfig: &container.HostConfig{RestartPolicy: opts.RestartPolicy}},
+	)
 }
 
 func setupDinD(t *testing.T) *dindContainer {