+++ pkg/provider/dockerswarm/events_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/moby/moby/api/types/swarm"
+	"github.com/moby/moby/client"
 	"github.com/neilotoole/slogt"
 	"github.com/sablierapp/sablier/pkg/provider/dockerswarm"
 	"gotest.tools/v3/assert"
@@ -29,13 +29,14 @@ func TestDockerSwarmProvider_NotifyInstanceStopped(t *testing.T) {
 	go p.NotifyInstanceStopped(ctx, waitC)
 
 	t.Run("service is scaled to 0 replicas", func(t *testing.T) {
-		service, _, err := dind.client.ServiceInspectWithRaw(ctx, c.ID, swarm.ServiceInspectOptions{})
+		inspectResult, err := dind.client.ServiceInspect(ctx, c.ID, client.ServiceInspectOptions{})
 		assert.NilError(t, err)
+		service := inspectResult.Service
 
 		replicas := uint64(0)
 		service.Spec.Mode.Replicated.Replicas = &replicas
 
-		_, err = p.Client.ServiceUpdate(ctx, service.ID, service.Version, service.Spec, swarm.ServiceUpdateOptions{})
+		_, err = p.Client.ServiceUpdate(ctx, service.ID, client.ServiceUpdateOptions{Version: service.Version, Spec: service.Spec})
 		assert.NilError(t, err)
 
 		name := <-waitC
@@ -45,10 +46,11 @@ func TestDockerSwarmProvider_NotifyInstanceStopped(t *testing.T) {
 	})
 
 	t.Run("service is removed", func(t *testing.T) {
-		service, _, err := dind.client.ServiceInspectWithRaw(ctx, c.ID, swarm.ServiceInspectOptions{})
+		inspectResult, err := dind.client.ServiceInspect(ctx, c.ID, client.ServiceInspectOptions{})
 		assert.NilError(t, err)
+		service := inspectResult.Service
 
-		err = p.Client.ServiceRemove(ctx, service.ID)
+		_, err = p.Client.ServiceRemove(ctx, service.ID, client.ServiceRemoveOptions{})
 		assert.NilError(t, err)
 
 		name := <-waitC