+++ pkg/provider/dockerswarm/service_inspect_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/moby/moby/api/types/container"
-	"github.com/moby/moby/api/types/swarm"
+	"github.com/moby/moby/client"
 	"github.com/neilotoole/slogt"
 	"github.com/sablierapp/sablier/pkg/provider/dockerswarm"
 	"github.com/sablierapp/sablier/pkg/sablier"
@@ -41,10 +41,11 @@ func TestDockerSwarmProvider_GetState(t *testing.T) {
 						return "", err
 					}
 
-					service, _, err := dind.client.ServiceInspectWithRaw(ctx, s.ID, swarm.ServiceInspectOptions{})
+					inspectResult, err := dind.client.ServiceInspect(ctx, s.ID, client.ServiceInspectOptions{})
 					if err != nil {
 						return "", err
 					}
+					service := inspectResult.Service
 
 					<-time.After(5 * time.Second)
 
@@ -77,10 +78,11 @@ func TestDockerSwarmProvider_GetState(t *testing.T) {
 						return "", err
 					}
 
-					service, _, err := dind.client.ServiceInspectWithRaw(ctx, s.ID, swarm.ServiceInspectOptions{})
+					inspectResult, err := dind.client.ServiceInspect(ctx, s.ID, client.ServiceInspectOptions{})
 					if err != nil {
 						return "", err
 					}
+					service := inspectResult.Service
 
 					return service.Spec.Name, nil
 				},
@@ -101,14 +103,15 @@ func TestDockerSwarmProvider_GetState(t *testing.T) {
 						return "", err
 					}
 
-					service, _, err := dind.client.ServiceInspectWithRaw(ctx, s.ID, swarm.ServiceInspectOptions{})
+					inspectResult, err := dind.client.ServiceInspect(ctx, s.ID, client.ServiceInspectOptions{})
 					if err != nil {
 						return "", err
 					}
+					service := inspectResult.Service
 
 					replicas := uint64(0)
 					service.Spec.Mode.Replicated.Replicas = &replicas
-					_, err = dind.client.ServiceUpdate(ctx, s.ID, service.Version, service.Spec, swarm.ServiceUpdateOptions{})
+					_, err = dind.client.ServiceUpdate(ctx, s.ID, client.ServiceUpdateOptions{Version: service.Version, Spec: service.Spec})
 					if err != nil {
 						return "", err
 					}