+++ pkg/provider/dockerswarm/testcontainers_test.go
@@ -25,29 +25,31 @@ type MimicOptions struct {
 	Labels        map[string]string
 }
 
-func (d *dindContainer) CreateMimic(ctx context.Context, opts MimicOptions) (swarm.ServiceCreateResponse, error) {
+func (d *dindContainer) CreateMimic(ctx context.Context, opts MimicOptions) (client.ServiceCreateResult, error) {
 	if len(opts.Cmd) == 0 {
 		opts.Cmd = []string{"/mimic", "-running", "-running-after=1s", "-healthy=false"}
 	}
 
 	d.t.Log("Creating mimic service with options", opts)
 	var replicas uint64 = 1
-	return d.client.ServiceCreate(ctx, swarm.ServiceSpec{
-		Mode: swarm.ServiceMode{
-			Replicated: &swarm.ReplicatedService{Replicas: &replicas},
-		},
-		TaskTemplate: swarm.TaskSpec{
-			RestartPolicy: opts.RestartPolicy,
-			ContainerSpec: &swarm.ContainerSpec{
-				Image:       "sablierapp/mimic:v0.3.1",
-				Healthcheck: opts.Healthcheck,
-				Command:     opts.Cmd,
+	return d.client.ServiceCreate(ctx, client.ServiceCreateOptions{
+		Spec: swarm.ServiceSpec{
+			Mode: swarm.ServiceMode{
+				Replicated: &swarm.ReplicatedService{Replicas: &replicas},
+			},
+			TaskTemplate: swarm.TaskSpec{
+				RestartPolicy: opts.RestartPolicy,
+				ContainerSpec: &swarm.ContainerSpec{
+					Image:       "sablierapp/mimic:v0.3.1",
+					Healthcheck: opts.Healthcheck,
+					Command:     opts.Cmd,
+				},
+			},
+			Annotations: swarm.Annotations{
+				Labels: opts.Labels,
 			},
 		},
-		Annotations: swarm.Annotations{
-			Labels: opts.Labels,
-		},
-	}, swarm.ServiceCreateOptions{})
+	})
 }
 
 func setupDinD(t *testing.T) *dindContainer {
@@ -73,7 +75,7 @@ func setupDinD(t *testing.T) *dindContainer {
 	// assert.NilError(t, err)
 
 	// Initialize the swarm
-	_, err = dindCli.SwarmInit(ctx, swarm.InitRequest{
+	_, err = dindCli.SwarmInit(ctx, client.SwarmInitOptions{
 		ListenAddr: "0.0.0.0",
 	})
 	assert.NilError(t, err)